feat(auth): run account-page 2FA enrollment in a dialog instead of navigating to /auth
Enabling 2FA from the account page sent the user off to the full-page
/auth/two-factor-setup wizard, which is jarring for a settings action.
The flow now lives in useTwoFactorEnrollment, which returns the current
step's heading and body separately so each entry point can present it
natively: a full page for organization-mandated enrollment after
sign-in, a dialog in place on the account page. Both get the same order
and the same download-gated recovery codes from one implementation.

Abandoning the dialog part-way is safe: verified against a running
backend, enable() leaves two_factor_enabled false until a code is
verified, so an abandoned enrollment changes nothing about the account
and starting again replaces the pending row.
